+++ src/atspm/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.1.3"
+__version__ = "2.2.0"
 
 from .sample_data import sample_data
 from .signal_data_processor import SignalDataProcessor